requirements.txt
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
